Added match to the DB upon completion

+++ app/backend/src/imgtransfer/imgtransfer.service.ts
@@ -25,8 +25,7 @@ export class ImgTransferService {
       const imgPth = path
         .join(__dirname, "/img", path.basename(imageUrl))
         .replace("/dist/src/imgtransfer", "");
-        if (fs.existsSync(imgPth))
-          fs.unlinkSync(imgPth);
+      if (fs.existsSync(imgPth)) fs.unlinkSync(imgPth);
     }
     this.prismaService.updateAvatar(userName, data.URL);
     return;
